+++ config/taxonomy.php
@@ -1,8 +1,12 @@
 <?php
 
 return [
+    
     'models' => [
+        
         'term' => \Fomvasss\Taxonomy\Models\Term::class,
+        
         'vocabulary' => \Fomvasss\Taxonomy\Models\Vocabulary::class,
+        
     ],
 ];