+++ src/Models/Traits/HasTaxonomies.php
@@ -5,7 +5,7 @@ namespace Fomvasss\Taxonomy\Models\Traits;
 use Fomvasss\Taxonomy\Models\Term;
 
 /**
- * Трейт для классов-моделей с таксономией
+ * Трейт для пользовательских классов-моделей котории имею таксономию (статьи, товары,...)
  *
  * Trait HasTaxonomies
  *
@@ -15,7 +15,7 @@ trait HasTaxonomies
 {
     /**
      * Связь:
-     * Сущность текущей модели "относится" к разным термам
+     * Сущность текущей модели "относится" к разным термам.
      *
      * @return \Illuminate\Database\Eloquent\Relations\MorphToMany
      */
@@ -25,81 +25,79 @@ trait HasTaxonomies
 
         return $this->morphToMany($related, 'termable');
     }
-
+    
     /**
-     * Связь:
-     * Сущность текущая модель "имеет" один терм
-     *
-     * @param null $foreignKey
-     * @param null $ownerKey
-     * @return mixed
-     */
-    public function term($foreignKey = null, $ownerKey = null)
-    {
-        $related = config('taxonomy.models.term', Term::class);
-
-        return $this->belongsTo($related, $foreignKey, $ownerKey);
-    }
-
-    /**
-     * Связь текущей модели с термамы по указанному словарю (по умолчанию ключ словаря - system_name)
-     * Используется для создание связей с термамы нужного словаря в своих моделях
+     * Связь текущей модели с термамы по указанному system_name словаря.
+     * Используется для создание связей с термамы нужного словаря в пользовательских моделях.
      *
      * @param $vocabulary
      * @param string|null $vocabularyKey
      * @return mixed
      */
-    public function termsByVocabulary($vocabulary, string $vocabularyKey = 'system_name')
+    public function termsByVocabulary($vocabulary)
     {
-        return $this->terms()->whereHas('vocabulary', function ($q) use ($vocabulary, $vocabularyKey) {
-            $q->where($vocabularyKey, $vocabulary);
-        });
+        return $this->terms()->where('vocabulary', $vocabulary);
     }
 
     /**
-     * Термы текущей модели по указанному словарю.
+     * Термы текущей модели по указанному system_name словаря.
      *
      * @param $query
      * @param $vocabulary
      * @param string $vocabularyKey
      * @return \Illuminate\Database\Eloquent\Builder|static
      */
-    public function scopeTermsByVocabulary($query, $vocabulary, string $vocabularyKey = 'system_name')
+    public function scopeTermsByVocabulary($query, $vocabulary)
     {
-        return $query->whereHas('terms.vocabulary', function ($v) use ($vocabulary, $vocabularyKey) {
-            $v->where($vocabularyKey, $vocabulary);
+        return $query->whereHas('terms', function ($t) use ($vocabulary) {
+            $t->where('vocabulary', $vocabulary);
         });
     }
 
     /**
      * Сущности по указанным термам с соответствующих
      * указанных словарей, например:
-     * получить все статьи терма-категории "WEB-Программирование" с словаря "Категории"
+     * получить все Статьи терма-категории "WEB-Программирование" (23) с словаря "Категории статтей" И 
+     * терма-города "Киев" (35) или "Москва" (56) с словаря "Города"
      *
      * @param $query
-     * @param array $taxonomies Например (для наглядности в примере slug-и):
+     * @param array $taxonomies Например:
      *  [
-     *      'regions' => ['moscow', 'rostov'],
-     *      'categories' => 'auto-remont'
+     *      'article_categories' => 23,
+     *      'cities' => [35, 56]
      *  ]
      * @param string $termKey
      * @param string $vocabularyKey
      * @return mixed
      */
-    public function scopeByTaxonomies($query, array $taxonomies, string $termKey = 'id', string $vocabularyKey = 'system_name')
+    public function scopeByTaxonomies($query, array $taxonomies, string $termKey = 'id')
     {
         foreach ($taxonomies as $vocabulary => $terms) {
             $terms = is_array($terms) ? $terms : [$terms];
             if (! empty($terms)) {
-                $query->whereHas('terms', function ($t) use ($vocabulary, $terms, $termKey, $vocabularyKey) {
-//                    $t->where('type', $vocabulary);
-                    $t->whereHas('vocabulary', function ($v) use ($vocabulary, $terms, $termKey, $vocabularyKey) {
-                        $v->where($vocabularyKey, $vocabulary);
-                    })->whereIn($termKey, $terms);
+                $query->whereHas('terms', function ($t) use ($vocabulary, $terms, $termKey) {
+                    $t->where('vocabulary', $vocabulary)->whereIn($termKey, $terms);
                 });
             }
         }
 
         return $query;
+    }
+
+
+    /**
+     * Связь:
+     * Сущность текущая модель "имеет" один терм.
+     * Ключ для связи находится в таблице сущности (Ex.: articles.term_id)
+     *
+     * @param null $foreignKey
+     * @param null $ownerKey
+     * @return mixed
+     */
+    public function term($foreignKey = null, $ownerKey = null, $relation = null)
+    {
+        $related = config('taxonomy.models.term', Term::class);
+
+        return $this->belongsTo($related, $foreignKey, $ownerKey, $relation);
     }
 }