+++ src/Models/Traits/HasTaxonomyablesToMany.php
@@ -8,7 +8,7 @@ use Fomvasss\Taxonomy\Models\Vocabulary;
 /**
  * Сущность текущей модели "относится" к разным
  * словарям или термам
- * (для класса модели к которому подключен трейт - используется не для категоризации)
+ * (для класса модели к которому подключен трейт - методы не используются для категоризации)
  *
  * Trait HasTaxonomyablesToMany
  *
@@ -17,26 +17,28 @@ use Fomvasss\Taxonomy\Models\Vocabulary;
 trait HasTaxonomyablesToMany
 {
     /**
-     * Сущность текущей модели "относится" к разным словарям
-     * Получить к список словарей, к которым относится
+     * Сущность текущей модели "относится" к разным словарям.
+     * Получить к список словарей, к которым относится.
      *
      * @return $this
      */
     public function vocabulariesToMany()
     {
         $related = config('taxonomy.models.vocabulary', Vocabulary::class);
+        
         return $this->morphToMany($related, 'vocabularyable');
     }
 
     /**
-     * Сущность текущей модели "относится" к разным термам
-     * Получить список термов, к которым относится
+     * Сущность текущей модели "относится" к разным термам.
+     * Получить список термов, к которым относится.
      *
      * @return $this
      */
     public function termsToMany()
     {
         $related = config('taxonomy.models.term', Term::class);
+        
         return $this->morphToMany($related, 'termable');
     }
 }