+++ src/Models/Vocabulary.php
@@ -16,39 +16,42 @@ class Vocabulary extends Model
 
     /**
      * Связь:
-     * Словарь имет много термов
+     * Словарь имеет много термов
      *
      * @return \Illuminate\Database\Eloquent\Relations\HasMany
      */
     public function terms()
     {
         $related = config('taxonomy.models.term', Term::class);
-        return $this->hasMany($related);
+        
+        return $this->hasMany($related, 'vocabulary');
     }
 
     /**
      * Связь:
-     * Сущность текущей модели (словарь) "держит" много термов
-     * (полиморфная связь! - не используется в категоризации)
+     * Сущность текущей модели - словарь "держит" много термов
+     * (полиморфная связь! - не используется в этом пакете для категоризации)
      *
      * @return $this
      */
     public function termsByMany()
     {
         $related = config('taxonomy.models.term', Term::class);
+        
         return $this->morphedByMany($related, 'vocabularyable');
     }
 
     /**
      * Связь:
-     * Сущность текущей модели "держит" много словарей
-     * (полиморфная связь! - не используется в категоризации)
+     * Сущность текущей модели - словарь "держит" много словарей
+     * (полиморфная связь! - не используется в этом пакете для категоризации)
      *
      * @return $this
      */
     public function vocabulariesByMany()
     {
-        $related = config('taxonomy.models.vocabulary', Vocabulary::class);
+        $related = config('taxonomy.models.vocabulary', static::class);
+        
         return $this->morphedByMany($related, 'vocabularyable');
     }
 }